chore(suite-common): Resolve circular dependency in trading reducer
