+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -70,19 +70,6 @@ packages/suite/src/views/onboarding/steps/SelectBackupType/SelectBackupType.tsx 
 packages/transport/src/transports/abstract.ts > packages/transport/src/transports/bridge.ts
 suite-common/graph/src/graphDataFetching.ts > suite-common/graph/src/graphUtils.ts
 suite-common/intl-types/src/index.ts > suite-common/intl-types/src/types.ts > packages/suite/src/components/suite/Translation.tsx
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/buy/index.ts > suite-common/trading/src/thunks/buy/confirmBuyTradeThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/buy/index.ts > suite-common/trading/src/thunks/buy/confirmBuyTradeThunk.ts > suite-common/trading/src/selectors/tradingSelectors.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/buy/index.ts > suite-common/trading/src/thunks/buy/handleBuyRequestThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/confirmExchangeTradeThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/sendDexTransactionThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/sendDexTransactionThunk.ts > suite-common/trading/src/thunks/common/index.ts > suite-common/trading/src/thunks/common/loadInitialDataThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/sendDexTransactionThunk.ts > suite-common/trading/src/thunks/common/index.ts > suite-common/trading/src/thunks/common/verifyAddressThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/sendDexTransactionThunk.ts > suite-common/trading/src/thunks/common/index.ts > suite-common/trading/src/thunks/common/watchTradeThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/sendTransactionThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/exchange/index.ts > suite-common/trading/src/thunks/exchange/signDataAndConfirmThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/sell/index.ts > suite-common/trading/src/thunks/sell/confirmSellTradeThunk.ts > suite-common/trading/src/thunks/sell/handleSellTradeThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/sell/index.ts > suite-common/trading/src/thunks/sell/handleSellRequestThunk.ts
-suite-common/trading/src/reducers/tradingReducer.ts > suite-common/trading/src/thunks/sell/index.ts > suite-common/trading/src/thunks/sell/sendSellTransactionThunk.ts
 suite-common/wallet-core/src/index.ts > suite-common/wallet-core/src/device/deviceThunks.ts > suite-common/bluetooth/src/index.ts > suite-common/bluetooth/src/bluetoothReducer.ts
 suite-common/wallet-core/src/index.ts > suite-common/wallet-core/src/device/deviceThunks.ts > suite-common/thp/src/index.ts > suite-common/thp/src/autoInitThpAfterDeviceConnectionThunk.ts
 suite-common/wallet-core/src/index.ts > suite-common/wallet-core/src/device/deviceThunks.ts > suite-common/thp/src/index.ts > suite-common/thp/src/finishThpAutoconnectThunk.ts
