+++ suite-common/trading/src/hooks/useSelector.ts
@@ -1,6 +1,6 @@
 import { TypedUseSelectorHook, shallowEqual, useSelector as useReduxSelector } from 'react-redux';
 
-import type { TradingRootState } from '../reducers/tradingReducer';
+import type { TradingRootState } from '../reducers/tradingCommonReducer';
 
 /**
  * Properly typed useSelector hook, use this one instead of directly importing it from react-redux.
