+++ suite-common/trading/src/reducers/tradingReducer.ts
@@ -1,166 +1,33 @@
 import { PayloadAction } from '@reduxjs/toolkit';
-import { Coins, CryptoId, InfoResponse, Platforms } from 'invity-api';
 
 import { createSliceWithExtraDeps } from '@suite-common/redux-utils';
-import { AccountKey, PrecomposedTransactionFinal } from '@suite-common/wallet-types';
-import { CardanoOutput, FeeLevel, PROTO } from '@trezor/connect';
 
+import {
+    TRADING_BUY_PREFIX,
+    TRADING_EXCHANGE_PREFIX,
+    TRADING_EXTENDED_PREFIX,
+    TRADING_PREFIX,
+    TRADING_SELL_PREFIX,
+    TRADING_SETTINGS_PREFIX,
+} from '../constants';
 import { buyThunks } from '../thunks/buy';
 import { exchangeThunks } from '../thunks/exchange';
 import { sellThunks } from '../thunks/sell';
-import {
-    TradingPaymentMethodListProps,
-    TradingTransaction,
-    TradingType,
-    TradingVerifiedAddress,
-} from '../types';
-import { TradingBuyState, buyInitialState, tradingBuyReducer } from './buyReducer';
-import { TRADING_PREFIX, TRADING_SETTINGS_PREFIX } from '../constants';
-import {
-    TradingExchangeState,
-    exchangeInitialState,
-    tradingExchangeReducer,
-} from './exchangeReducer';
-import { TradingSellState, sellInitialState, tradingSellReducer } from './sellReducer';
-import {
-    TradingSettingsState,
-    settingsInitialState,
-    tradingSettingsReducer,
-} from './settingsReducer';
-
-type TradingComposedTransactionInfoOutputs = {
-    outputs?: PROTO.TxOutputType[] | CardanoOutput[];
-};
-
-export interface TradingComposedTransactionInfo {
-    composed?: Pick<
-        PrecomposedTransactionFinal,
-        | 'fee'
-        | 'feePerByte'
-        | 'feeLimit'
-        | 'estimatedFeeLimit'
-        | 'maxFeePerGas'
-        | 'maxPriorityFeePerGas'
-        | 'token'
-    > &
-        TradingComposedTransactionInfoOutputs;
-    selectedFee?: FeeLevel['label'];
-}
-
-export interface TradingInfo {
-    platforms?: Platforms;
-    coins?: Coins;
-    paymentMethods: TradingPaymentMethodListProps[];
-}
-
-export interface TradingPrefilledFromAccount {
-    cryptoId: CryptoId | undefined;
-    key: AccountKey | undefined;
-}
-
-export interface TradingState {
-    info: TradingInfo;
-    buy: TradingBuyState;
-    exchange: TradingExchangeState;
-    sell: TradingSellState;
-    composedTransactionInfo: TradingComposedTransactionInfo;
-    trades: TradingTransaction[];
-    modalCryptoId: CryptoId | undefined;
-    modalAccountKey: AccountKey | undefined;
-    isLoading: boolean;
-    lastLoadedTimestamp: number;
-    activeSection: TradingType;
-    prefilledFromAccount: TradingPrefilledFromAccount;
-    verifiedAddress: TradingVerifiedAddress;
-    settings: TradingSettingsState;
-}
-
-export type TradingRootState = {
-    wallet: {
-        trading: TradingState;
-    };
-};
-
-export const initialState: TradingState = {
-    info: {
-        platforms: undefined,
-        coins: undefined,
-        paymentMethods: [],
-    },
-    buy: buyInitialState,
-    exchange: exchangeInitialState,
-    sell: sellInitialState,
-    composedTransactionInfo: {},
-    trades: [],
-    isLoading: false,
-    modalAccountKey: undefined,
-    modalCryptoId: undefined,
-    lastLoadedTimestamp: 0,
-    activeSection: 'buy',
-    prefilledFromAccount: {
-        cryptoId: undefined,
-        key: undefined,
-    },
-    verifiedAddress: undefined,
-    settings: settingsInitialState,
-};
+import { TradingTransaction } from '../types';
+import { tradingBuyReducer } from './buyReducer';
+import { tradingExchangeReducer } from './exchangeReducer';
+import { tradingSellReducer } from './sellReducer';
+import { tradingSettingsReducer } from './settingsReducer';
+import { initialState, tradingCommonReducer } from './tradingCommonReducer';
 
 type StorageActionPayload = {
     tradingTrades?: TradingTransaction[];
 };
 
-export const tradingSlice = createSliceWithExtraDeps({
-    name: TRADING_PREFIX,
+const tradingSlice = createSliceWithExtraDeps({
+    name: TRADING_EXTENDED_PREFIX,
     initialState,
-    reducers: {
-        saveInfo(state, action: PayloadAction<InfoResponse>) {
-            state.info.coins = action.payload.coins;
-            state.info.platforms = action.payload.platforms;
-        },
-        savePaymentMethods(state, action: PayloadAction<TradingPaymentMethodListProps[]>) {
-            state.info.paymentMethods = action.payload;
-        },
-        saveComposedTransactionInfo(state, action: PayloadAction<TradingComposedTransactionInfo>) {
-            state.composedTransactionInfo = action.payload;
-        },
-        saveTrade(state, action: PayloadAction<TradingTransaction>) {
-            if (action.payload.key) {
-                const trades = state.trades.filter(t => t.key !== action.payload.key);
-                trades.push(action.payload);
-
-                state.trades = trades;
-            }
-        },
-        setModalCryptoCurrency(state, action: PayloadAction<CryptoId | undefined>) {
-            state.modalCryptoId = action.payload;
-        },
-        setModalAccountKey(state, action: PayloadAction<string | undefined>) {
-            state.modalAccountKey = action.payload;
-        },
-        setLoading(
-            state,
-            action: PayloadAction<{ isLoading: boolean; lastLoadedTimestamp?: number }>,
-        ) {
-            state.isLoading = action.payload.isLoading;
-            state.lastLoadedTimestamp = action.payload.lastLoadedTimestamp ?? 0;
-        },
-        setTradingActiveSection(state, action: PayloadAction<TradingType>) {
-            state.activeSection = action.payload;
-        },
-        setTradingFromPrefilledAccount(
-            state,
-            action: PayloadAction<{
-                cryptoId: CryptoId | undefined;
-                key: string | undefined;
-            }>,
-        ) {
-            state.prefilledFromAccount.cryptoId = action.payload.cryptoId;
-            state.prefilledFromAccount.key = action.payload.key;
-        },
-        setVerifiedAddress(state, action: PayloadAction<TradingVerifiedAddress>) {
-            state.verifiedAddress = action.payload;
-        },
-    },
+    reducers: {},
     extraReducers: (builder, extra) => {
         builder
             .addCase(
@@ -220,13 +87,26 @@ export const tradingSlice = createSliceWithExtraDeps({
                     tradingSettingsReducer(state.settings, action);
                 },
             )
-            .addDefaultCase((state, action) => {
-                tradingBuyReducer(state.buy, action);
-                tradingSellReducer(state.sell, action);
-                tradingExchangeReducer(state.exchange, action);
-            });
+            .addMatcher(
+                action => action.type.startsWith(TRADING_BUY_PREFIX),
+                (state, action) => {
+                    tradingBuyReducer(state.buy, action);
+                },
+            )
+            .addMatcher(
+                action => action.type.startsWith(TRADING_EXCHANGE_PREFIX),
+                (state, action) => {
+                    tradingExchangeReducer(state.exchange, action);
+                },
+            )
+            .addMatcher(
+                action => action.type.startsWith(TRADING_SELL_PREFIX),
+                (state, action) => {
+                    tradingSellReducer(state.sell, action);
+                },
+            )
+            .addMatcher(action => action.type.startsWith(TRADING_PREFIX), tradingCommonReducer);
     },
 });
 
 export const prepareTradingReducer = tradingSlice.prepareReducer;
-export const tradingActions = tradingSlice.actions;
