+++ suite-common/trading/src/selectors/__tests__/settingsSelectors.test.ts
@@ -1,4 +1,4 @@
-import { type TradingRootState, initialState } from '../../reducers/tradingReducer';
+import { type TradingRootState, initialState } from '../../reducers/tradingCommonReducer';
 import { selectTradingMaxSlippagePercentage } from '../settingsSelectors';
 
 describe('settingsSelectors', () => {
