+++ suite-common/trading/src/thunks/buy/__tests__/selectBuyQuoteThunk.test.ts
@@ -13,7 +13,8 @@ import { configureMockStore, extraDependenciesMock } from '@suite-common/test-ut
 import { MIN_MAX_QUOTES_OK } from '../../../__fixtures__/buyUtils';
 import { invityAPI } from '../../../invityAPI';
 import { BuyInfo, TradingBuyState } from '../../../reducers/buyReducer';
-import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
+import { initialState } from '../../../reducers/tradingCommonReducer';
+import { prepareTradingReducer } from '../../../reducers/tradingReducer';
 import { TradingCountryCode } from '../../../types';
 import { buyThunks } from '../index';
 import { SelectBuyQuoteThunkProps } from '../selectBuyQuoteThunk';
