+++ suite-common/trading/src/thunks/buy/confirmBuyTradeThunk.ts
@@ -7,7 +7,7 @@ import { Account } from '@suite-common/wallet-types';
 import { TRADING_BUY_THUNK_PREFIX } from '../../constants';
 import { invityAPI } from '../../invityAPI';
 import { tradingBuyActions } from '../../reducers/buyReducer';
-import { tradingActions } from '../../reducers/tradingReducer';
+import { tradingActions } from '../../reducers/tradingCommonReducer';
 import {
     selectTradingBuyReceiveAccountKey,
     selectTradingBuySelectedQuote,
