+++ suite-common/trading/src/thunks/exchange/__tests__/sendDexTransactionThunk.test.ts
@@ -8,7 +8,8 @@ import { Account } from '@suite-common/wallet-types';
 import { MIN_MAX_QUOTES_OK } from '../../../__fixtures__/exchangeUtils';
 import { accountBtc } from '../../../__fixtures__/utils';
 import { TradingExchangeState } from '../../../reducers/exchangeReducer';
-import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
+import { initialState } from '../../../reducers/tradingCommonReducer';
+import { prepareTradingReducer } from '../../../reducers/tradingReducer';
 import { tradingThunks } from '../../common';
 import * as confirmExchangeTradeThunk from '../confirmExchangeTradeThunk';
 import { exchangeThunks } from '../index';
