+++ suite-common/trading/src/thunks/exchange/__tests__/sendTransactionThunk.test.ts
@@ -9,7 +9,8 @@ import { Account } from '@suite-common/wallet-types';
 import { MIN_MAX_QUOTES_OK } from '../../../__fixtures__/exchangeUtils';
 import { accountBtc } from '../../../__fixtures__/utils';
 import { TradingExchangeState } from '../../../reducers/exchangeReducer';
-import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
+import { initialState } from '../../../reducers/tradingCommonReducer';
+import { prepareTradingReducer } from '../../../reducers/tradingReducer';
 import { TradingTransactionExchange } from '../../../types';
 import { tradingThunks } from '../../common';
 import { exchangeThunks } from '../index';
