+++ suite-common/trading/src/thunks/exchange/signDataAndConfirmThunk.ts
@@ -10,7 +10,7 @@ import { transformTypedData } from '@trezor/connect-plugin-ethereum';
 
 import { confirmExchangeTradeThunk } from './confirmExchangeTradeThunk';
 import { TRADING_EXCHANGE_THUNK_PREFIX } from '../../constants';
-import { tradingActions } from '../../reducers/tradingReducer';
+import { tradingActions } from '../../reducers/tradingCommonReducer';
 import {
     selectTradingExchangeAccountKey,
     selectTradingExchangeReceiveAccountKey,
