+++ suite-common/trading/src/thunks/sell/__tests__/sendSellTransactionThunk.test.ts
@@ -9,7 +9,8 @@ import { Account } from '@suite-common/wallet-types';
 import { accountBtc } from '../../../__fixtures__/utils';
 import { invityAPI } from '../../../invityAPI';
 import { TradingSellState, sellInitialState } from '../../../reducers/sellReducer';
-import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
+import { initialState } from '../../../reducers/tradingCommonReducer';
+import { prepareTradingReducer } from '../../../reducers/tradingReducer';
 import { TradingSellFormProps, TradingTransactionSell } from '../../../types';
 import { sellUtilsFixtures } from '../../../utils/sell/__fixtures__/sellUtils';
 import { tradingThunks } from '../../common';
