+++ suite-common/trading/src/thunks/sell/handleSellRequestThunk.ts
@@ -7,7 +7,7 @@ import { convertAmountSubunitsToUnits } from '@suite-common/wallet-utils';
 import { TRADING_DEFAULT_SELL_FLOWS, TRADING_SELL_THUNK_PREFIX } from '../../constants';
 import { invityAPI } from '../../invityAPI';
 import { tradingSellActions } from '../../reducers/sellReducer';
-import { tradingActions } from '../../reducers/tradingReducer';
+import { tradingActions } from '../../reducers/tradingCommonReducer';
 import { selectTradingCoinSymbolByCryptoId } from '../../selectors/tradingSelectors';
 import { HandleSellRequestThunkProps, MinimalSellFormProps, TradingSellType } from '../../types';
 import {
